Implement a camera and simple functionality for player & dog.
The camera is setup appropriately, using a programmatic approach to ensure that we have flexibilty.
The human and dog are implemented to have basic movement functionality. They should move in unison and
maintain a gap. They don't reflect when they move left so the movement is janky at the moment.