The main focus of this commit is treat throwing
First, I made the treat land where the mouse cursor is to make a more interactive
feature. Second, the treat is sped up a bit forcing the dog to chase it in this
iteration. Third, if the dog detects the treat, it will fixate on it and chase it.
Before there was a jerky motion as the dog wouldn't detect a treat every loop.